Add files via upload
Project name: Auburn Tigers Documentation 

Objective: This project was developed with the purpose of recognizing an NCAA basketball team, the Auburn Tigers. This project, visualized by JavaFx, and SceneBuilder, documents the players of the whole team. When the user runs the program, they are introduced to a page where all of the players and laid out. And when the user preses on one, they are directed to a new page, which shortly describes their backgrounds, statistics on games played, points scored on average, and more. 

Attributes: Error handling, JavaFX.