Add files via upload
Project name: Digital Photo Frame 

Objective: The purpose of this project is to make a visual display of 5 pictures that commemorates the GOAT of the UFC 29-0 Khabib Nurmagomedov. The program passively switches to the next picture every 3 seconds, but the user is able to immediately skip to the next by clicking the skip button. 

Attributes: JavaFx, SceneBuilder